FIX: added casting to prevent "signed/unsigned mismatch" warning in Visual Studio

+++ src/core/f-dtoa.c
@@ -3576,7 +3576,7 @@ rv_alloc(int i)
 
 	j = sizeof(ULong);
 	for(k = 0;
-		sizeof(Bigint) - sizeof(ULong) - sizeof(int) + j <= i;
+		sizeof(Bigint) - sizeof(ULong) - sizeof(int) + j <= (unsigned int)i;
 		j <<= 1)
 			k++;
 	r = (int*)Balloc(k);
